# Conflicts:
#	include/etl/generators/type_traits_generator.h
#	include/etl/profiles/arduino_arm.h
#	include/etl/profiles/armv5.h
#	include/etl/profiles/armv5_no_stl.h
#	include/etl/profiles/armv6.h
#	include/etl/profiles/armv6_no_stl.h
#	include/etl/profiles/armv7.h
#	include/etl/profiles/armv7_no_stl.h
#	include/etl/profiles/cpp03.h
#	include/etl/profiles/cpp03_no_stl.h
#	include/etl/profiles/cpp11.h
#	include/etl/profiles/cpp11_no_stl.h
#	include/etl/profiles/cpp14.h
#	include/etl/profiles/cpp14_no_stl.h
#	include/etl/profiles/cpp17.h
#	include/etl/profiles/cpp17_no_stl.h
#	include/etl/profiles/determine_builtin_support.h
#	include/etl/profiles/ticc.h
#	include/etl/profiles/ticc_no_stl.h
#	include/etl/type_traits.h

+++ include/etl/private/chrono/operators.h
@@ -38,291 +38,255 @@ namespace etl
   {
     //*************************************************************************
     // month_day
-    inline ETL_CONSTEXPR14 etl::chrono::month_day operator /(const etl::chrono::month& m,
-                                                             const etl::chrono::day& d) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_day operator/(const etl::chrono::month& m, const etl::chrono::day& d) ETL_NOEXCEPT
     {
       return etl::chrono::month_day(m, d);
     }
 
     // month_day
-    inline ETL_CONSTEXPR14 etl::chrono::month_day operator /(const etl::chrono::month& m, 
-                                                             int d) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_day operator/(const etl::chrono::month& m, int d) ETL_NOEXCEPT
     {
-      return etl::chrono::month_day(m, etl::chrono::day(d));
+      return etl::chrono::month_day(m, etl::chrono::day(static_cast<unsigned>(d)));
     }
 
     // month_day
-    inline ETL_CONSTEXPR14 etl::chrono::month_day operator /(int m, 
-                                                             const etl::chrono::day& d) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_day operator/(int m, const etl::chrono::day& d) ETL_NOEXCEPT
     {
-      return etl::chrono::month_day(etl::chrono::month(m), d);
+      return etl::chrono::month_day(etl::chrono::month(static_cast<unsigned>(m)), d);
     }
 
     // month_day
-    inline ETL_CONSTEXPR14 etl::chrono::month_day operator /(const etl::chrono::day& d, 
-                                                             const etl::chrono::month& m) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_day operator/(const etl::chrono::day& d, const etl::chrono::month& m) ETL_NOEXCEPT
     {
       return etl::chrono::month_day(m, d);
     }
 
     // month_day
-    inline ETL_CONSTEXPR14 etl::chrono::month_day operator /(const etl::chrono::day& d, 
-                                                             int m) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_day operator/(const etl::chrono::day& d, int m) ETL_NOEXCEPT
     {
-      return etl::chrono::month_day(etl::chrono::month(m), d);
+      return etl::chrono::month_day(etl::chrono::month(static_cast<unsigned>(m)), d);
     }
 
     //*************************************************************************
     // month_day_last
-    inline ETL_CONSTEXPR14 etl::chrono::month_day_last operator /(const etl::chrono::month& m,
-                                                                  etl::chrono::last_spec) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_day_last operator/(const etl::chrono::month& m, etl::chrono::last_spec) ETL_NOEXCEPT
     {
       return etl::chrono::month_day_last(m);
     }
 
     // month_day_last
-    inline ETL_CONSTEXPR14 etl::chrono::month_day_last operator /(int m, 
-                                                                  etl::chrono::last_spec) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_day_last operator/(int m, etl::chrono::last_spec) ETL_NOEXCEPT
     {
-      return etl::chrono::month_day_last(etl::chrono::month(m));
+      return etl::chrono::month_day_last(etl::chrono::month(static_cast<unsigned>(m)));
     }
 
     // month_day_last
-    inline ETL_CONSTEXPR14 etl::chrono::month_day_last operator /(etl::chrono::last_spec,
-                                                                  const etl::chrono::month& m) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_day_last operator/(etl::chrono::last_spec, const etl::chrono::month& m) ETL_NOEXCEPT
     {
       return etl::chrono::month_day_last(m);
     }
-     
+
     // month_day_last
-    inline ETL_CONSTEXPR14 etl::chrono::month_day_last operator/(etl::chrono::last_spec, 
-                                                                 int m) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_day_last operator/(etl::chrono::last_spec, int m) ETL_NOEXCEPT
     {
-      return etl::chrono::month_day_last(etl::chrono::month(m));
+      return etl::chrono::month_day_last(etl::chrono::month(static_cast<unsigned>(m)));
     }
 
     //*************************************************************************
     // month_weekday
-    inline ETL_CONSTEXPR14 etl::chrono::month_weekday operator /(const etl::chrono::month& m,
-                                                                 const etl::chrono::weekday_indexed& wdi) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_weekday operator/(const etl::chrono::month& m, const etl::chrono::weekday_indexed& wdi) ETL_NOEXCEPT
     {
       return etl::chrono::month_weekday(m, wdi);
     }
 
     // month_weekday
-    inline ETL_CONSTEXPR14 etl::chrono::month_weekday operator /(int m, 
-                                                                 const etl::chrono::weekday_indexed& wdi) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_weekday operator/(int m, const etl::chrono::weekday_indexed& wdi) ETL_NOEXCEPT
     {
-      return etl::chrono::month_weekday(etl::chrono::month(m), wdi);
+      return etl::chrono::month_weekday(etl::chrono::month(static_cast<unsigned>(m)), wdi);
     }
 
     // month_weekday
-    inline ETL_CONSTEXPR14 etl::chrono::month_weekday operator /(const etl::chrono::weekday_indexed& wdi,
-                                                                 const etl::chrono::month& m) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_weekday operator/(const etl::chrono::weekday_indexed& wdi, const etl::chrono::month& m) ETL_NOEXCEPT
     {
       return etl::chrono::month_weekday(m, wdi);
     }
 
     // month_weekday
-    inline ETL_CONSTEXPR14 etl::chrono::month_weekday operator /(const etl::chrono::weekday_indexed& wdi, 
-                                                                 int m) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_weekday operator/(const etl::chrono::weekday_indexed& wdi, int m) ETL_NOEXCEPT
     {
-      return etl::chrono::month_weekday(etl::chrono::month(m), wdi);
+      return etl::chrono::month_weekday(etl::chrono::month(static_cast<unsigned>(m)), wdi);
     }
 
     //*************************************************************************
     // month_weekday_last
-    inline ETL_CONSTEXPR14 etl::chrono::month_weekday_last operator /(const etl::chrono::month& m,
-                                                                      const etl::chrono::weekday_last& wdl) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_weekday_last operator/(const etl::chrono::month& m, const etl::chrono::weekday_last& wdl) ETL_NOEXCEPT
     {
       return etl::chrono::month_weekday_last(m, wdl);
     }
 
     // month_weekday_last
-    inline ETL_CONSTEXPR14 etl::chrono::month_weekday_last operator /(int m,
-                                                                      const etl::chrono::weekday_last& wdl) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_weekday_last operator/(int m, const etl::chrono::weekday_last& wdl) ETL_NOEXCEPT
     {
-      return etl::chrono::month_weekday_last(etl::chrono::month(m), wdl);
+      return etl::chrono::month_weekday_last(etl::chrono::month(static_cast<unsigned>(m)), wdl);
     }
 
     // month_weekday_last
-    inline ETL_CONSTEXPR14 etl::chrono::month_weekday_last operator /(const etl::chrono::weekday_last& wdl,
-                                                                      const etl::chrono::month& m) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_weekday_last operator/(const etl::chrono::weekday_last& wdl, const etl::chrono::month& m) ETL_NOEXCEPT
     {
       return etl::chrono::month_weekday_last(m, wdl);
     }
 
     // month_weekday_last
-    inline ETL_CONSTEXPR14 etl::chrono::month_weekday_last operator /(const etl::chrono::weekday_last& wdl,
-                                                                      int m) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::month_weekday_last operator/(const etl::chrono::weekday_last& wdl, int m) ETL_NOEXCEPT
     {
-      return etl::chrono::month_weekday_last(etl::chrono::month(m), wdl);
+      return etl::chrono::month_weekday_last(etl::chrono::month(static_cast<unsigned>(m)), wdl);
     }
 
     //*************************************************************************
     // year_month
-    inline ETL_CONSTEXPR14 etl::chrono::year_month operator /(const etl::chrono::year& y,
-                                                              const etl::chrono::month& m) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month operator/(const etl::chrono::year& y, const etl::chrono::month& m) ETL_NOEXCEPT
     {
       return etl::chrono::year_month(y, m);
     }
 
     // year_month
-    inline ETL_CONSTEXPR14 etl::chrono::year_month operator /(const etl::chrono::year& y,
-                                                              int m) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month operator/(const etl::chrono::year& y, int m) ETL_NOEXCEPT
     {
-      return etl::chrono::year_month(y, etl::chrono::month(m));
+      return etl::chrono::year_month(y, etl::chrono::month(static_cast<unsigned>(m)));
     }
 
     ////*************************************************************************
     // year_month_day
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator /(const etl::chrono::year_month& ym,
-                                                                  const etl::chrono::day& d) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator/(const etl::chrono::year_month& ym, const etl::chrono::day& d) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_day(ym.year(), ym.month(), d);
     }
 
     // year_month_day
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator /(const etl::chrono::year_month& ym, 
-                                                                  int d ) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator/(const etl::chrono::year_month& ym, int d) ETL_NOEXCEPT
     {
-      return etl::chrono::year_month_day(ym.year(), ym.month(), etl::chrono::day(d));
+      return etl::chrono::year_month_day(ym.year(), ym.month(), etl::chrono::day(static_cast<unsigned>(d)));
     }
 
     // year_month_day
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator /(const etl::chrono::year& y,
-                                                                  const etl::chrono::month_day& md) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator/(const etl::chrono::year& y, const etl::chrono::month_day& md) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_day(y, md.month(), md.day());
     }
 
     // year_month_day
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator/(int y,
-                                                                 const etl::chrono::month_day& md) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator/(int y, const etl::chrono::month_day& md) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_day(etl::chrono::year(y), md.month(), md.day());
     }
 
     // year_month_day
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator /(const etl::chrono::month_day& md,
-                                                                  const etl::chrono::year& y) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator/(const etl::chrono::month_day& md, const etl::chrono::year& y) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_day(y, md.month(), md.day());
     }
 
     // year_month_day
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator /(const etl::chrono::month_day& md,
-                                                                  int y) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_day operator/(const etl::chrono::month_day& md, int y) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_day(etl::chrono::year(y), md.month(), md.day());
     }
 
     //*************************************************************************
     // year_month_day_last
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_day_last operator /(const etl::chrono::year_month& ym,
-                                                                       etl::chrono::last_spec) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_day_last operator/(const etl::chrono::year_month& ym, etl::chrono::last_spec) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_day_last(ym.year(), etl::chrono::month_day_last(ym.month()));
     }
 
     // year_month_day_last
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_day_last operator /(const etl::chrono::year& y,
-                                                                       const etl::chrono::month_day_last& mdl) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_day_last operator/(const etl::chrono::year& y, const etl::chrono::month_day_last& mdl) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_day_last(y, mdl);
     }
 
     // year_month_day_last
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_day_last operator /(int y,
-                                                                       const etl::chrono::month_day_last& mdl) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_day_last operator/(int y, const etl::chrono::month_day_last& mdl) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_day_last(etl::chrono::year(y), mdl);
     }
 
     // year_month_day_last
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_day_last operator /(const etl::chrono::month_day_last& mdl,
-                                                                       const etl::chrono::year& y) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_day_last operator/(const etl::chrono::month_day_last& mdl, const etl::chrono::year& y) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_day_last(y, mdl);
     }
 
     // year_month_day_last
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_day_last operator /(const etl::chrono::month_day_last& mdl,
-                                                                       int y) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_day_last operator/(const etl::chrono::month_day_last& mdl, int y) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_day_last(etl::chrono::year(y), mdl);
     }
 
     //*************************************************************************
     // year_month_weekday
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday operator /(const etl::chrono::year_month& ym,
-                                                                      const etl::chrono::weekday_indexed& wdi) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday operator/(const etl::chrono::year_month& ym, const etl::chrono::weekday_indexed& wdi)
+      ETL_NOEXCEPT
     {
       return etl::chrono::year_month_weekday(ym.year(), ym.month(), wdi);
     }
 
     // year_month_weekday
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday operator /(const etl::chrono::year& y,
-                                                                      const etl::chrono::month_weekday& mwd) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday operator/(const etl::chrono::year& y, const etl::chrono::month_weekday& mwd) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_weekday(y, mwd.month(), mwd.weekday_indexed());
     }
 
     // year_month_weekday
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday operator/(int y, 
-                                                                     const etl::chrono::month_weekday& mwd) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday operator/(int y, const etl::chrono::month_weekday& mwd) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_weekday(etl::chrono::year(y), mwd.month(), mwd.weekday_indexed());
     }
 
     // year_month_weekday
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday operator /(const etl::chrono::month_weekday& mwd,
-                                                                      const etl::chrono::year& y) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday operator/(const etl::chrono::month_weekday& mwd, const etl::chrono::year& y) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_weekday(y, mwd.month(), mwd.weekday_indexed());
     }
 
     // year_month_weekday
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday operator /(const etl::chrono::month_weekday& mwd, 
-                                                                      int y) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday operator/(const etl::chrono::month_weekday& mwd, int y) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_weekday(etl::chrono::year(y), mwd.month(), mwd.weekday_indexed());
     }
 
     //*************************************************************************
     // year_month_weekday_last
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday_last operator /(const etl::chrono::year_month& ym,
-                                                                           const etl::chrono::weekday_last& wdl) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday_last operator/(const etl::chrono::year_month& ym, const etl::chrono::weekday_last& wdl)
+      ETL_NOEXCEPT
     {
       return etl::chrono::year_month_weekday_last(ym.year(), ym.month(), wdl);
     }
 
     // year_month_weekday_last
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday_last operator /(const etl::chrono::year& y,
-                                                                           const etl::chrono::month_weekday_last& mwdl) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday_last operator/(const etl::chrono::year& y, const etl::chrono::month_weekday_last& mwdl)
+      ETL_NOEXCEPT
     {
       return etl::chrono::year_month_weekday_last(y, mwdl.month(), mwdl.weekday_last());
     }
 
     // year_month_weekday_last
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday_last operator/(int y,
-                                                                          const etl::chrono::month_weekday_last& mwdl) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday_last operator/(int y, const etl::chrono::month_weekday_last& mwdl) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_weekday_last(etl::chrono::year(y), mwdl.month(), mwdl.weekday_last());
     }
 
     // year_month_weekday_last
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday_last operator /(const etl::chrono::month_weekday_last& mwdl,
-                                                                           const etl::chrono::year& y) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday_last operator/(const etl::chrono::month_weekday_last& mwdl, const etl::chrono::year& y)
+      ETL_NOEXCEPT
     {
       return etl::chrono::year_month_weekday_last(y, mwdl.month(), mwdl.weekday_last());
     }
 
     // year_month_weekday_last
-    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday_last operator /(const etl::chrono::month_weekday_last& mwdl,
-                                                                           int y) ETL_NOEXCEPT
+    inline ETL_CONSTEXPR14 etl::chrono::year_month_weekday_last operator/(const etl::chrono::month_weekday_last& mwdl, int y) ETL_NOEXCEPT
     {
       return etl::chrono::year_month_weekday_last(etl::chrono::year(y), mwdl.month(), mwdl.weekday_last());
     }
-  }
-}
+  } // namespace chrono
+} // namespace etl
