+++ include/etl/scaled_rounding.h
@@ -31,9 +31,9 @@ SOFTWARE.
 #ifndef ETL_SCALED_ROUNDING_INCLUDED
 #define ETL_SCALED_ROUNDING_INCLUDED
 
-#include "static_assert.h"
-#include "type_traits.h"
+#include "platform.h"
 #include "absolute.h"
+#include "type_traits.h"
 
 namespace etl
 {
@@ -63,9 +63,7 @@ namespace etl
   /// \return Unscaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_ceiling_unscaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_ceiling_unscaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -90,9 +88,7 @@ namespace etl
   /// \return Scaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_ceiling_scaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_ceiling_scaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -105,9 +101,7 @@ namespace etl
   /// \return Unscaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_floor_unscaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_floor_unscaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
     if (Scaling == 1)
@@ -131,9 +125,7 @@ namespace etl
   /// \return Scaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_floor_scaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_floor_scaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -147,9 +139,7 @@ namespace etl
   /// \return Unscaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_half_up_unscaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_half_up_unscaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -177,9 +167,7 @@ namespace etl
   /// \return Scaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_half_up_scaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_half_up_scaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -193,9 +181,7 @@ namespace etl
   /// \return Unscaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_half_down_unscaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_half_down_unscaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -221,9 +207,7 @@ namespace etl
   /// \return Scaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_half_down_scaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_half_down_scaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -236,9 +220,7 @@ namespace etl
   /// \return Unscaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_zero_unscaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_zero_unscaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -258,9 +240,7 @@ namespace etl
   /// \return Scaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_zero_scaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_zero_scaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -273,9 +253,7 @@ namespace etl
   /// \return Unscaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_infinity_unscaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_infinity_unscaled(T value) ETL_NOEXCEPT
   {
     if (value >= 0)
     {
@@ -293,9 +271,7 @@ namespace etl
   /// \return Scaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_infinity_scaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_infinity_scaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -309,9 +285,7 @@ namespace etl
   /// \return Unscaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_half_even_unscaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_half_even_unscaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -325,7 +299,7 @@ namespace etl
       if ((etl::absolute(value) % scale_t(Scaling)) == scale_t(Scaling / 2U))
       {
         // Odd?
-        if ((value / scale_t(Scaling)) & 1U)
+        if (static_cast<unsigned int>(value / scale_t(Scaling)) & 1U)
         {
           return T(round_half_up_unscaled<Scaling>(value));
         }
@@ -348,9 +322,7 @@ namespace etl
   /// \return Scaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_half_even_scaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_half_even_scaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -364,9 +336,7 @@ namespace etl
   /// \return Unscaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_half_odd_unscaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_half_odd_unscaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
@@ -380,7 +350,7 @@ namespace etl
       if ((etl::absolute(value) % scale_t(Scaling)) == scale_t(Scaling / 2U))
       {
         // Odd?
-        if ((value / scale_t(Scaling)) & 1U)
+        if (static_cast<unsigned int>(value / scale_t(Scaling)) & 1U)
         {
           return T(round_half_down_unscaled<Scaling>(value));
         }
@@ -403,14 +373,12 @@ namespace etl
   /// \return Scaled, rounded integral.
   //***************************************************************************
   template <uint32_t Scaling, typename T>
-  ETL_NODISCARD
-  ETL_CONSTEXPR14
-  T round_half_odd_scaled(T value) ETL_NOEXCEPT
+  ETL_NODISCARD ETL_CONSTEXPR14 T round_half_odd_scaled(T value) ETL_NOEXCEPT
   {
     typedef typename scaled_rounding_t<T>::type scale_t;
 
     return T(round_half_odd_unscaled<Scaling>(value) * scale_t(Scaling));
   }
-}
+} // namespace etl
 
 #endif
